Create README.md
Android application that was developed for the foi mobile app challenge where it had won the competition. The application enables the user to provide feedback for a certain topic that was asked from the administrator. User is able to send feedback in form of a picture,video,rating and a specific comment which are sent to a web application and analyzed.
User is also able to chose different groups(like sports, reading, history) which helps administrator to select and filter users for the desired reports.